refactor (test/directives): ts'ify test/directives
Translate all of the AtScript code in .../test/directives to TypeScript.
